+++ R/add-flag.R
@@ -25,10 +25,10 @@ add_flags <- function(data, y_var, index, flag){
     #add flag
     flags <- data[[coln]]
 
-    flags[index] <- unlist(lapply(flags[index], function(old_flag) {
+    flags[index] <- lapply(flags[index], function(old_flag) {
       old_flag <- unlist(old_flag)
       sort(unique(na.omit(c(old_flag, flag))))
-    }))
+    })
 
     data[[coln]] <- flags
   }