+++ tests/testthat/test-step2-explore-plot.R
@@ -88,6 +88,7 @@ test_that("{shinytest2} recording: checking-module2", {
   app$expect_values(export = "data2-table", name="datasum-table-weekly",screenshot_args = FALSE) #cal check
 
   #test reverting changes
+  app$set_inputs(`data2-table_opt` = "Change Log") #have to put back to changelog to let it work
   app$set_inputs(`data2-log_table_rows_selected` = 4, allow_no_input_binding_ = TRUE)
   app$set_inputs(`data2-conf` = TRUE, allow_no_input_binding_ = TRUE, priority_ = "event")
   app$set_inputs(`data2-table_opt` = "Change Log")