+++ NEWS.md
@@ -6,13 +6,29 @@
 
 ### Bug Fixes
 
--   Can no longer select rows in the other table options within check data module. This allowed you to unintentially view past data versions.
+-   Can no longer select rows in the other table options within check data module. This allowed you to unintentionally view past data versions.
 
 -   Bug in interpolation code to fill in flags was causing the first flag of the data to apply to the entire dataset.
 
-### **Updates**
+-   Fixed a number of bugs with how colors were displaying in the plot, particularly when using file colors along with other color schemes.
+
+-   Selected points now clears in the quality flags module when switching the method.
+
+-   App no longer crashes if precipitation data can't be fetched.
+
+-   Issue with flags added to existing flags was fixed.
+
+### Updates
 
 -   Now in the quality flags module, visualizing those flags starts selected so you don't have to remember to select it to view the flags you applied after flagging.
+
+-   File names for saving files now clear when data is cleared.
+
+-   Metadata files can now recognize missing values even if they're not quite how R expects them.
+
+-   Site coordinates are now stored if specified before the project is loaded.
+
+-   Exported the `get_nldas` function.
 
 ## SondePolishR 0.0.9004